+++ tests/examples/predictie_2019_allfromdia.py
@@ -31,9 +31,7 @@ stats_no4Y = []
 stats_no19Y = []
 
 for current_station in selected_stations:
-    print('-'*50)
-    print('%-45s = %s'%('station_name',current_station))
-    print('-'*5)
+    print(f'current_station: {current_station}')
     
     # station settings
     nodalfactors = True