+++ tests/examples/predictie_2019_frommergedcomp.py
@@ -24,9 +24,7 @@ selected_stations = ['CADZD','DORDT','ABDN']#,'AUKFPFM']
 stats_noana = []
 
 for current_station in selected_stations:
-    print('-'*50)
-    print('%-45s = %s'%('station_name',current_station))
-    print('-'*5)
+    print(f'current_station: {current_station}')
     
     # station settings
     nodalfactors = True